Add initial version with defaults for Penbu chat
The Windows Classic window generator creates a window in Windows
Classic style drawn in SVG. The user can specify the width and height
of the resulting image as well as the scale at which to draw the window
elements.

This was requested by Penbu for a chat window overlay. The defaults
follow his use case specifically right now, but will be changed to
something more generic in the future.

Features:
- Preview
- Download as SVG and PNG
- Set image height, width, and scale
- Toggle scrollbar, status bar, resize gripper, and maximized style
- Reset to default