Make ResizeObserver callback cross-browser
It appears that the type of contentBoxSize changed from object to array
at some point in the last three years or so. Pale Moon still implements
the object form, but Firefox and likely the other major browsers
implement the array form. To support both, the array is unwrapped in
the array case.