Clarify that the output SVG is not covered under the project licence.
Output is not typically covered under the project's licence, but the SVG
file is significantly contained within the project source code, putting
it in scope of the licence by default. Explicitly exclude the SVG string
from the scope of the licence to preserve normal expectations.

See also: The Bison exception to the GPL